Migrate from Makefile to Grunt build process
In order to simplify the dependencies for this project and move towards a grunt based build convention that has gained popularity and offers portial     future flexibility the Makefile-based build process on this template has been changed to a Grunt-based process.

The grunt-based process features all of the functionality of the previous Makefile:
  - setup: copies the sample local_settings file from "local_settings.py.sample" to "local_settings.py" to make it active
  - compass: uses compass as a preprocessor for sass
  - cssmin: minifies the processed css
  - requirejs: handles javascript dependencies (build config moved from /static/scripts/javascript/app.build.js to gruntfile.
  - coffee: compiles coffeescript to js
  - watch: watches scss and coffee files and live processes them if any changes are detected.
  - collect-static: runs Djangos collectstatic utility to symlink files

By default the grunt file will run the "build" task which runs:
  - setup
  - coffee
  - compass
  - cssmin
  - requirejs

This commit also adds a "package.json" file to handle dependencies for Grunt. Its important that "npm install" is run prior to using the Gruntfile